backend/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
blis==0.7.11
catalogue==2.0.10
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.3
cryptography==41.0.5
ctransformers==0.2.27
cupy-cuda11x==12.2.0
cymem==2.0.8
dataclasses-json==0.6.2
distro==1.8.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl#sha256=6215d71a3212690e9aec49408a27e3fe6ad7cd6c715476e93d70dc784041e93e
faiss-cpu==1.7.4
fastapi==0.104.1
fastrlock==0.8.2
filelock==3.13.1
frozenlist==1.4.0
fsspec==2023.10.0
future==0.18.3
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.1
huggingface-hub==0.17.3
idna==3.4
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.334
langcodes==3.3.0
langsmith==0.0.63
MarkupSafe==2.1.3
marshmallow==3.20.1
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.10
mypy-extensions==1.0.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.1
openai==1.2.3
packaging==23.2
pdfminer.six==20221105
Pillow==10.1.0
pke @ git+https://github.com/boudinfl/pke.git@69871ffdb720b83df23684fea53ec8776fd87e63
preshed==3.0.9
py-cpuinfo==9.0.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.11.3
sentence-transformers==2.2.2
sentencepiece==0.1.99
smart-open==6.4.0
sniffio==1.3.0
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.23
srsly==2.4.8
starlette==0.27.0
sympy==1.12
tenacity==8.2.3
thinc==8.2.1
threadpoolctl==3.2.0
tiktoken==0.5.1
tokenizers==0.14.1
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1
transformers==4.35.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.8.0
Unidecode==1.3.7
urllib3==2.0.7
uvicorn==0.24.0.post1
wasabi==1.1.2
watchfiles==0.21.0
weasel==0.3.4
websockets==12.0
yarl==1.9.2
